% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_deconvolution_palette.R
\name{add_deconvolution_palette}
\alias{add_deconvolution_palette}
\title{Supply distinctive colours to a graph}
\usage{
add_deconvolution_palette(gg_plot, return_palette = FALSE)
}
\arguments{
\item{gg_plot}{A ggplot graph.}

\item{return_palette}{Logical, TRUE will return the palette only, FALSE (the default) will return the modified graph.}
}
\description{
Supply distinctive colours to a graph
}
