% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration_curve.R
\name{plot_calibration_curve}
\alias{plot_calibration_curve}
\title{Plot calibration curve}
\usage{
plot_calibration_curve(x, y, scale_in, scale_out, ..., intercept = TRUE)
}
\arguments{
\item{x}{numeric, the ground truth (spiked) proportion/ratio values.}

\item{y}{numeric, estimated mixing proportions/ratios.}

\item{scale_in, scale_out}{character, possible values: \code{"proportion"} or
\code{"ratio"}. Decide whether input should be transformed from proportions
to ratios and vice versa.}

\item{...}{expression(s) (don't use ” or "", just quote or rlang::expr) that
will be evaluated in order to modify the plot (e.g. add new geom, change
axis labels etc.).}

\item{intercept}{logical, should the fitted calibration line include an
intercept?}
}
\description{
Plot calibration curve
}
\examples{
if (!require(dplyr)) {
  library(dplyr)
}
dat <- results_MS_spectrum \%>\%
  filter(z == 9 & spectrum_type == "imp") \%>\%
  distinct(ground_truth, scaling_factor)
plot_calibration_curve(x = dat$ground_truth, y = dat$scaling_factor, scale_in = "proportion",
scale_out = "proportion")
}
