% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimate_bias.R
\name{plot_estimate_bias}
\alias{plot_estimate_bias}
\title{Visualise Bias of Estimates}
\usage{
plot_estimate_bias(x, y, error_type, ..., replicates = FALSE, smooth = FALSE)
}
\arguments{
\item{x}{numeric, the ground truth (spiked) proportion/ratio values.}

\item{y}{vector or matrix with estimates (columns).}

\item{error_type}{character, \code{"diff"} shows y-x, \code{"relative diff"}
100*(y-x)/x, \code{"multiplicative err"} y/x}

\item{...}{to be documented}

\item{replicates}{logical, set to TRUE if there are replicate spectra per
given spiked-in proportion; as a result, points will be drawn instead of
lines.}

\item{smooth}{logical, add smoothed line using \code{ggplot2::geom_smooth}?}
}
\description{
Create scatterplot presenting difference between estimates and ground truth,
on absolute, relative, or multiplicatve scale.
}
\examples{
#' if (!require(tidyverse)) {
  library(tidyverse)
}
# prepare data
dat <- results_MS_spectrum \%>\%
  filter(spectrum_type == "imp")
ground_truth <- dat \%>\% distinct(sample, ground_truth) \%>\% pull(ground_truth)
dat <- dat \%>\%
  distinct(sample, z, ground_truth, scaling_factor) \%>\%
  pivot_wider(id = sample, names_from = "z", values_from = scaling_factor, names_prefix = "z") \%>\%
  select(-sample)

# "absolute error", replicated concentrations, smoother line
plot_estimate_bias(x = ground_truth, y = dat, error_type = "diff", replicates = TRUE, smooth = TRUE)
}
