% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectrum_ms.R
\name{read_spectrum_ms}
\alias{read_spectrum_ms}
\title{Read in MS spectrum}
\usage{
read_spectrum_ms(path, mode, do_norm = FALSE)
}
\arguments{
\item{path}{A character string of length 1 containing input file path.}

\item{mode}{"stick" or "profile" depending on the input data.}

\item{do_norm}{If TRUE, sum-to-one normalization will be applied.}
}
\value{
Returns a tibble with two columns: mz and intensity.
}
\description{
Read in single MS spectrum from a .csv (stick data) or .mzXML (profile data) file.
}
\examples{
path <- system.file("extdata", "raw_MS_spectrum.csv", package = "deconvolutionutils")
head(read_spectrum_ms(path, mode = "stick"))
}
