% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_MS_spectrum.R
\docType{data}
\name{results_MS_spectrum}
\alias{results_MS_spectrum}
\title{Deconvolution results data set created to showcase how to call various plotting functions}
\format{
A data frame (in long format) with 1392 rows and 7 variables used as example for plot_spectrum() function.
It contains the following variables: mz, intensity, spectrum_type, z (i.e. charge state), sample, ground_truth, scaling_factor.
}
\usage{
results_MS_spectrum
}
\description{
Deconvolution results data set created to showcase how to call various plotting functions
}
\keyword{datasets}
