% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimate_bias.R
\name{plot_estimate_bias}
\alias{plot_estimate_bias}
\title{Visualise Bias of Estimates}
\usage{
plot_estimate_bias(
  data,
  x,
  y,
  error_type,
  ...,
  replicates = FALSE,
  smooth = FALSE,
  extra_ellipsis = NULL
)
}
\arguments{
\item{data}{Data in longer format containing the mapping variables.}

\item{x}{Bare variable name of the ground truth, i.e. spiked-in proportion.}

\item{y}{Bare variable name of the estimated proportions.}

\item{error_type}{character, \code{"diff"} shows y-x, \code{"relative diff"}
100*(y-x)/x, \code{"multiplicative err"} y/x}

\item{...}{Pass additional mappings to \code{ggplot} like colour, symbol, size etc. using bare
variable names.}

\item{replicates}{logical, set to TRUE if there are replicate spectra per
given spiked-in proportion; as a result, points will be drawn instead of
lines.}

\item{smooth}{logical, add smoothed line using \code{ggplot2::geom_smooth}?}

\item{extra_ellipsis}{expression(s) (don't use ” or "", just quote or
rlang::expr) that will be evaluated in order to modify the plot (e.g. add
new geom, change axis labels etc.).}
}
\description{
Create scatterplot presenting difference between estimates and ground truth,
on absolute, relative, or multiplicatve scale.
}
\examples{
if (!require(tidyverse)) {
  library(tidyverse)
}
dat <- results_MS_spectrum \%>\%
  filter(spectrum_type == "imp") \%>\%
  distinct(sample, z, ground_truth, scaling_factor) \%>\%
  mutate(z = as.factor(z))
# "absolute error", replicated concentrations, smoother line
plot_estimate_bias(data = dat, x = ground_truth, y = scaling_factor, error_type = "diff", colour = z, replicates = TRUE, smooth = TRUE)
}
