% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma.R
\name{plot_ma}
\alias{plot_ma}
\title{MA plot of estimates vs ground truth}
\usage{
plot_ma(x, y, ...)
}
\arguments{
\item{x}{numeric, the ground truth (spiked) proportion/ratio values.}

\item{y}{numeric, estimated mixing proportions/ratios.}

\item{...}{expression(s) (don't use ” or "", just quote or rlang::expr) that
will be evaluated in order to modify the plot (e.g. add new geom, change
axis labels etc.).}
}
\description{
MA plot of estimates vs ground truth
}
\examples{
if (!require(tidyverse)) {
  library(tidyverse)
}
dat <- results_MS_spectrum \%>\%
  filter(z == 9 & spectrum_type == "imp") \%>\%
  distinct(ground_truth, scaling_factor)
plot_ma(x = dat$ground_truth, y = dat$scaling_factor)
}
